Add results and visualisation for AL on toy problem.
Also now only consider inequality constraints to be violated (for
updating penalty parameter) when they take on a value > epsilon, rather
than 0 as before. This is because otherwise the penalty parameter gets
too small, which causes difficulties when optimising the resulting AL.

Perhaps it would be better to instead enforce a lower bound on the
penalty parameter.