Add updated optimiser for AL acquisition function
Added a refactored optimiser for the AL acquisition function. With this
optimiser, when performing L-BFGS-B we force one of the starting points
to be the most recently queried point. This is an attempt to solve the
issue of the optimiser failing once the penaty parameter in the AL
becomes large, which is an issue which arises intermittently.