Increase robustness of L-BFGS-B optimisation of Augmented Lagrangian
In a few runs, once the penalty parameter became very small this caused
the L-BFGS-B optimisation of the acquisition function to fail. I suspect
this is due to the gradient at the constraint satisfaction boundary
becoming very large.

If this happens, we now stop reducing the penalty parameter, and
increase it until the acquisition function once again becomes amenable
to being optimised with the L-BFGS optimiser. Have also added a
MAX_OPTIMISATION_ATTEMPTS number to put a limit on the number of times
we try increasing the penalty parameter.