Updated KKT acqusition functions (both standard and TS)
Updated acquisition functions to avoid issue where we constantly favour
invalid points due to the EI term in the acqusition function. Now we
have the option to apply a mask so that we only have to consider points
in the feasible region. I suspect this will improve results.

Also added some visualisations of TS-KKT method in action.