Add TS-AL implementation fully consistent with Picheny et al's paper
Noticed that the AL value used for updating the penalty term/Lagrange
multipliers differed slightly between the main body and the appendix of
Picheny et al's paper, so added an implementation which is fully
consistent with the appendix and ran experiments to see if there was any
difference. Intuiutively I believe the rule in the appendix makes
slightly more sense, although both are very similar.

As suspected, performance difference was negligible.