Add functionality to run Lockwood problem
Added ability to run the Lockwood simulation and return results in a
manner which fits cleanly in the Trieste framework.

Also added a 'DeterministicAugmentedLagrangian' for investigating the
L-BFGS-B optimiser without GPs, to try and see why optimisation
sometimes fails. It still sometimes fails with deterministic functions,
so I suspect that the issue is indeed to do with the penalty term
becoming too high. Had a look at gradients calculated and they seemed to
be fine in the deterministic case.